// bun server with mongo db object id/src/schemas/store/store_carousel.schema.ts
export type CarouselItem = {
  _id: number;
  index: number;
  images: string;                      // ÃƒÂ¢Ã¢â‚¬Â Ã‚Â single image string
  imageAlt?: string;
  title?: string;
  titleColor?: string;
  tileBackground?: string;
  description?: string;
  descriptionColor?: string;
  discriptionBackgroundColor?: string;
  buttonText?: string;
  buttonTextColor?: string;
  buttonHoverTextColor?: string;
  buttonBackground?: string;
  buttonHoverBackground?: string;
  buttonShadow?: boolean;
  buttonShadowColor?: string;
  buttonBorder?: boolean;
  buttonBorderColor?: string;
  productId?: number;
  fontFamily?: string[];
  category?: string;
  overlayOpacity?: number;
  createdAt: string;
  updatedAt: string;
};

export const store_carousel = pgTable("store_carousel", {
  // _id: uuid("_id").defaultRandom().primaryKey(),
  _id: bigint("_id", { mode: "number" }).primaryKey().generatedAlwaysAsIdentity({ startWith: 17417 }),
  storeId: uuid("store_id")
    .notNull()
    .unique()
    .references(() => createStore._id, { onDelete: "cascade" }),
  carousels: jsonb("carousels")
    .$type<CarouselItem[]>()
    .notNull()
    .default([]),
  createdAt: timestamp("created_at").defaultNow().notNull(),
  updatedAt: timestamp("updated_at").defaultNow().notNull()
    .$onUpdate(() => new Date()),
}, (table) => ({
  storeIdx: index("store_carousel_store_idx").on(table.storeId),
}));